#49 add statistics to project
