--- app/assets/stylesheets/application.scss
@@ -1,1 +1,0 @@
-@import "node_modules/bootstrap/scss/bootstrap";
