--- app/assets/stylesheets/teams.scss
@@ -1,18 +1,0 @@
-// Place all the styles related to the teams controller here.
-// They will automatically be included in application.css.
-// You can use Sass (SCSS) here: https://sass-lang.com/
-
-.teams-cont{
-
-  display: flex;
-  justify-content: center;
-  div.teams {
-    ul {
-      padding-left: 0;
-    };
-    text-align:center;
-    background-color:gray;
-    width: 50%;
-    
-  }
-}
