+++ app/javascript/packs/application.js
@@ -9,8 +9,7 @@ require("@rails/activestorage").start()
 require("channels")
 require("packs/d3.v3.min")
 require("packs/d3.tip.v0.6.3")
-require("packs/custom/user_show")
-
+require("packs/custom/stats_show")
 
 // Uncomment to copy all static images under ../images to the output folder and reference
 // them with the image_pack_tag helper in views (e.g <%= image_pack_tag 'rails.png' %>)
