+++ app/javascript/stylesheets/application.scss
@@ -1,1 +1,3 @@
 @import "node_modules/bootstrap/scss/bootstrap";
+@import "teams";
+@import "stats_show";
