+++ app/views/layouts/_navbar.html.erb
@@ -5,7 +5,7 @@
       <% if current_user.present?%>
         <a class="nav-item nav-link active" data-turbolinks="false" href="/profile">Profile<span class="sr-only">(current)</span></a>
         <a class="nav-item nav-link" style="color:white" href="/teams">Team</a>
-        <a class="nav-item nav-link" style="color:white" href="/projects">Project</a>
+        <a class="nav-item nav-link" data-turbolinks="false" style="color:white" href="/projects">Project</a>
         <a class="nav-item nav-link" style="color:white" href="/tracks">Track</a>
         <div style="background-color: gray">
           <a class="nav-item nav-link" style="color:black" data-method="delete" href="/users/sign_out">Sign Out</a>
