+++ app/views/projects/index.html.erb
@@ -20,7 +20,7 @@
         <td><%= project.repository %></td>
         <td><%= project.cost %></td>
         <td><%= project.start_at %></td>
-        <td><%= link_to 'Show', project %></td>
+        <td data-turbolinks="false"><%= link_to 'Show', project %></td>
         <td><%= link_to 'Edit', edit_project_path(project) %></td>
         <td><%= link_to 'Destroy', project, method: :delete, data: { confirm: 'Are you sure?' } %></td>
       </tr>
